Python-generated schedules (just mirroring pipeline bundles) (#997)
Invoke either as `build/bin/tpp-sched` or `python -m mlir.tpp.sched`
(which requires `PYTHONPATH` to point to `build/python_packages` - the
`tools/tpp-sched/tpp-sched.py` script (installed as `tpp-sched`) takes
care of that).

When `tpp-sched` is invoked it just dumps the default pipeline schedule
to stdout. When a path is provided as a positional argument, i.e.
`tpp-sched.py PAYLOAD_FILE`, the schedule gets generated and immediately
applied to the payload and only the transformed payload gets dumped to
stdout. Use `--bundles=BUNDLE1,...,BUNDLEN` to control which bundles
make up the schedule. All the bundles' arguments available on `tpp-opt`
are available on `tpp-sched`.

When invoked as `tpp-sched --payload PAYLOAD_FILE`, it will just print
the payload and schedule together as one IR dump. This is meant to
facilitate working directly with the IR before passing it on to, e.g.,
`tpp-opt` (to apply the schedule to the payload use `build/bin/tpp-sched
--payload PAYLOAD_FILE | build/bin/tpp-opt --load-tpp-dialects
--transform-interpreter` -- as to why we need `--load-tpp-dialects`, see
[here](https://discourse.llvm.org/t/dynamic-pass-scheduling-and-dialect-loading/85166/2)).

Introduces Python-bindings for our own dialects, for now mostly to be
able to register them from Python before running passes that dependent
on them.

Requires MLIR to be build with `-DMLIR_ENABLE_BINDINGS_PYTHON=1`.

Adds RUN lines which use `tpp-sched` to each test which used `tpp-opt`
to apply a pipeline bundle.

For now we only have partial support for the GPU pipeline bundles.