Support lowering of mixed precision vector.contract (bf16xbf16->fp32) to AMX.
This patch support mixed precision lowering of vector.contract with bf16 input
and fp32 output type while also refactoring the code for readability/modularity
and also prepare for other possible mixed precisison operation.

It was observed that the 'REQUIRES' directive was not working due to a
lit config gap. This also fixes the observed issue.