Support lowering of mixed precision vector.contract to AMX. (#1070)
This patch support mixed precision lowering of vector.contract with
bf16/i8 input and fp32/i32 output type respectively
while also refactoring the code for readability/modularity. Moreover, it
also prepares for other possible mixed precision
operation.

It was observed that the 'REQUIRES' directive was not working due to a
lit config gap. This also fixes the observed issue.